// src/pages/types.ts
export interface Book {
  id: string;
  title: string;
  pdfUrl: string;
  bookUrl: string; // Updated to match the field name in Firebase
  coverUrl: string;
  type: string;
  description: string;
  language: string;
  category: string;
  price: number;
  currency: string;
  year: number;
  audioUrl?: string; 

  // Add other properties if needed
    // Add other properties if needed
  }
  